v. 3.0.0a-preview1
This is a preview of the next major update for the mod. The project was rewritten from scratch to fix some of the main issues and introduce support for the permissions API that you can use with mods like LuckPerms. The full changelog will be available when the final version of this update releases.